requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssmin==0.2.0
Flask==2.2.2
Flask-Assets==2.1.0
greenlet==3.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.6
Naked==0.1.31
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
pymysql-pooling==1.0.6
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
urllib3==1.26.12
webassets==2.0
Werkzeug==2.2.2
# 追加したライブラリ
Flask-SocketIO==5.3.7
eventlet==0.33.3
python-dotenv==1.0.0
